% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{example_function}
\alias{example_function}
\alias{cond_bad_argument}
\title{An example function}
\usage{
example_function(x)

cond_bad_argument(..., .call = getOption("cnd.call", TRUE))
}
\arguments{
\item{x}{A value}

\item{..., .call}{Condition information}
}
\description{
An example function
}
\section{Conditions}{


Conditions are generated through the \code{\link[cnd:cnd-package]{\{cnd\}}} package.
The following conditions are associated with this function:

\describe{

\item{\code{\link[=t0-cnd-conditions]{t0:bad_argument/error}}}{
x should be \code{TRUE}
}

}

For more conditions, see: \link{t0-cnd-conditions}
}

