% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t0-cnd-conditions.R
\name{t0-cnd-conditions}
\alias{t0-cnd-conditions}
\alias{t0:bad_argument}
\alias{bad_argument}
\alias{t0:bad_argument/error}
\title{Conditions for \code{t0}}
\description{
Conditions for \code{t0}
}
\details{
The following conditions are defined in the \code{{t0}} package.
}
\section{\code{{cnd}}}{

These conditions are made with the \code{\link[cnd:cnd-package]{\{cnd\}}} package though
the use of \code{\link[cnd:condition]{cnd::condition()}}.
}

\section{\code{{t0}} conditions}{


\subsection{\code{t0:bad_argument/error}}{
\describe{
\item{package}{\code{{t0}}}
\item{class}{\code{t0:bad_argument}}
\item{type}{\strong{error}}
}
x should be \code{TRUE}
}
}

\seealso{
\link[cnd:cnd-package]{cnd::cnd-package} \link[cnd:condition]{cnd::condition}
}
\keyword{internal}
